refactor(homepage): unify get page method
