+++ homepage/src/pages/cards.jsx
@@ -1,27 +1,14 @@
 import Head from 'next/head';
 import contentMapper from '../layouts/contentMapper';
-import { fetchPage } from '../lib/fetchPage';
-import { fetchCards } from '../lib/fetchCards';
-import { componentMapper } from '../lib/componentMapper';
-import { processCard } from '../lib/processCard';
 import { getLayout } from '../lib/getLayout';
+import { getPage } from '../lib/getPage';
 
 /**
  *
  * @type {import('next').GetStaticProps}
  */
 export const getStaticProps = async ({ locale }) => {
-  const rawCards = fetchCards(locale);
-  const cardTasks = rawCards.map(async (card) => {
-    return processCard(card, rawCards);
-  });
-  const cards = await Promise.all(cardTasks);
-
-  const page = fetchPage(locale, 'cards');
-
-  const contentList = page.data['layout_list']?.map((layout) =>
-    componentMapper(layout, cards),
-  );
+  const page = await getPage('cards', locale);
 
   const headInfo = {
     title: {
@@ -38,9 +25,7 @@ export const getStaticProps = async ({ locale }) => {
         title: headInfo.title[locale],
         description: '',
       },
-      page: {
-        contentList,
-      },
+      page,
       layout,
     },
   };
