+++ homepage/src/pages/resource.jsx
@@ -1,8 +1,7 @@
 import Head from 'next/head';
 import contentMapper from '../layouts/contentMapper';
-import { componentMapper } from '../lib/componentMapper';
-import { fetchPage } from '../lib/fetchPage';
 import { getLayout } from '../lib/getLayout';
+import { getPage } from '../lib/getPage';
 
 /**
  *
@@ -16,10 +15,7 @@ export const getStaticProps = async ({ locale }) => {
     },
   };
 
-  const page = fetchPage(locale, 'resource');
-  const contentList = page.data['layout_list']?.map((layout) =>
-    componentMapper(layout, []),
-  );
+  const page = await getPage('resource', locale);
 
   const layout = await getLayout(locale);
 
@@ -29,9 +25,7 @@ export const getStaticProps = async ({ locale }) => {
         title: headInfo.title[locale],
         description: '',
       },
-      page: {
-        contentList,
-      },
+      page,
       layout,
     },
   };
